Move (for now) ops package under hmb to easy packaging
The django project package, which we named ops should ideally not be
reusable, i.e. available on sys.path, but that means we 'd have to split
the installation directory, i.e. installing the app package under
dist-packages while the project package goes somewhater under
usr/share/hmb/ops but we don't know of an easy way to do that with
setuptools+dh-python just yet.