Implement more general canonical lookups
Canonical lookups can now be part of a chain of relation.  The
"canonical" keyword is not needed when looking up a relation, the
canonical name (the filter term) can be assigned directly to the foreign
field name.

Canonical lookup are now processed and turned into real lookups, that
Django understands, when Q objects are created.  This way we should be
able to use them almost anywhere, however these custom Q objects have to
be created explicitly.