Fix crash when sorting name/canonical columns
Somehow, when sorting, django-tables2 requests a len/count of a
values_list() which triggers a _fetch_all() call and the field list
given to values_list() contains 'canonical' which doesn't exist in the
database/is not a valid field name.  This patch simply removes
'canonical' from the retrieved fields, with no apparent downside (though
needs more testing).  A future implementation could replace "canonical"
with any real fieldnames needed to build the canonical name.