Centralize field/column name enumeration into models
The new Model.get_fields() returns a namedtuple with three lists,
listing the field object, the names, and the verbose names.  By default,
the auto id field is returned but not the man-to-many fields.  Models
from the healthy human cohort with Note fields get those returned
because there is usually only a single note attached to an object which
can be usually be displayed with no problems.  The code to hide id
columns when there is a name column present on the table page is
simplified.