Add base filter infrastructure
At this point this allows to set a filter on the default manager.  The
docs warn against using such filters on the default managers.  The patch
applied such filters to count columns.  Normal forward relations are not
filtered, so e.g. for the healthy human cohort database, we can filter
out non-consenting participants in the participants table but the fecal
sample table will still show samples from non-consenting participants.