models: fix confusion about which app defines Model.history / ChangeRecord
Before this patch, in an app deriving models from mibios.Model the
system check raises errors:

hhcd.ASV.history: (fields.E300) Field defines a relation with model 'ChangeRecord', which is either not installed, or is abstract.
hhcd.ASV.history: (fields.E307) The field hhcd.ASV.history was declared with a lazy reference to 'hhcd.changerecord', but app 'hhcd' doesn't provide model 'changerecord'.
hhcd.ASV_history.changerecord: (fields.E307) The field hhcd.ASV_history.changerecord was declared with a lazy reference to 'hhcd.changerecord', but app 'hhcd' doesn't provide model 'changerecord'.

Changing the order in which we define the models seems to fix this. We
also redo the migration/squash dance just to reduce the number of
migrations we have. Not sure why after the squash we got another
migration.