view: do not add reverse relation count column to average table
There seem to be joins created by these reverse relation counts
annotations that result in out average groups to have records multiple
times (one for each reverse relation).  No solution for this is
apparent, but thwe count column are not as critical for averaged tables,
one can easily go to the regular table to get those.