views: Ensure DatasetMixin users have all attributes set with defaults
This is the list of attributes set by in setup() so it should be in the
mixin not in the TableView class proper.  Setting these i the class
definition is mostly redundant since setup should set them in any case
but it's good to have a list well visible as reference.