Speed up count column totals calculation
Getting the count column footer totals requires going through all rows,
summing up the counts and not just for the displayed page.  So requires
an additional database access in any case.  This patch makes the totals
calculation the CountColumn call a new aggregate function that provides
just the totals for all count columns in one query.  The speedup
measured for the current ASV table (12000 rows) is 1.7 -> 1.3 seconds.