models: Re-implement making shared files correctly and faster
And use much less memory.  This is on-the-fly generation to be available
for download.  Nothing is kept in memory for other work.  The speedup
for a 650,000 non-zero count dataset is from 95 to 18 seconds and no
notable memory usage.  The old way used multiple gigabytes.

The old algorithm also seems to have produced wrong count numbers in
the final output but it's not yet clear why.