views/forms: support complex natural key search over multiple fields
Models need to pussort this by having their Model.natural_lookup()
return a dictionary when given a None argument.  The returned dict's
values should be None, and must be ignored.  The keys are used to to
refer to the (multiple) fields the natural key is derived from.  Drops
requirement to have a non-empty search term.