Offer normalized shared table download
There are two implemententations of normalizations.  The first is
relative abundance as fractions between 0 and 1.  The second is a
scaling of the counts to a normal sample size with home-made a minimal
maximum rounding error algorithm.  What we really want is something that
gets the manhattan vector of the given length with the smallest angle to
the original.