tables: Make table factory call signature more flexible
The patch allow the factory to be called with just a TableView as
argument. Verbose names will preferrably be obtained from the view and
obtaining them should now fail more gracefully.

It is now also possible to call the factory with just the model class
which may be useful for testing and debugging.