load: fix crashes in certain rare cases
This handles cases when a non-field column, maybe a composite
column/field is not handled by a parse_FOO() method remains, and gets
passed on towards the Model constructor parameters.  This patch prevents
them from being passed on to code that expects valid field names.

It's not clear if the second case (the blank validation stuff) can ever
happen.