models: re-organize abundance import code a bit
This possibly fixes issues with the import code running twice, but it's
not clear at all how the import ran a again (after what it seemed the
transaction finished) and how the patch fixes that.  Anyways, most of
the code now uses one fewer indentation level and the doc string has
instructions how to add the data.  Managers used are now the usual
objects not curated.