forms: Make parts of the export form re-usable
Instead of having the factory make all of the export forms, we now have
a abstract class hierarchy of a parent class with just the format
options and a child class that adds the field choices.  The factory sets
the dynamic attributes and adds the hidden state input.  The class
constructors inject the dynamic attributes into the form fields.