models: Switch to generate shared data from relative abundance
With this patch we allow the groups column values to be set by the
caller via a passed mapping function.  Multiple such column can be
specified.

Mothur-compatibility-mode is not yet impemented.

Normalization is implemented via simply multiplying the floating-point
abundances with the normal sample size.  Thus the sample size varies
somewhat.  The previous scaling implementetation is still in the code
but not used at the moment.