settings: streamline settings organization
Remove local and production modules.  Only a single settings module per
app.  An app depending on another will typically have the other's
settings module imported into their setting.  An instance of the site,
be it development or deployment should have an edited version of the
example settings available to be the instance's settings module.