views: Reduce number of queries for table view (in some cases)
Include relations in query if they are needed to generate a natural key.
Tis may save per-row queries when displaying the table at the cost of
one or more joins.  This currently affects the mibios_seq.Sequencing
table and adds joins with FecalSample and Participant to make
Sample.natural.  In this the main query actually takes more time to come
back than what is saved when avoiding for two small queries per row for
the natural key  (so we might end up reverting this at some point)