dataconfig: Support complex searches on numeric fields
Supports the following syntax. Te basic terms, where N and M are any
numeric string that is understood by the respective field:

    >=N >N <N <=N N-M N- -N - N,M

These terms can be combined with commas, the meaning of such
concatenation is the logical OR. -N is equivalent to >=N, and similarly
N- and <= mean the same. The single dash - means the whole range and
does not add any filtering.  No further optimization is performed, so
e.g. <5 and 1, 2, <5 will result in the same data set and the redundant
1, 2 will be part of the Q object passed to the ORM layer.  Spaces are
allowed and ignored, but the lexical processing is rather primitive, so
e.g. the search term "1 2" is equivalent with 12.