data: fix duplicate rows in some tables with complext filtering
When to appliy distinct() to a query set should be further explored.
An option is to always apply distinct() as the cost when it's not needed
should be small and so far there is no need for having query require
running without distinct().  With this patch only filters and excludes
are analysed.  Q-object are ignored so far.