views: Setup DataConfig in View.setup() not get()
The first patch in a series of improvements of the view code.  Expect
improved modularization, better logic how the views are set up, a new
view testing facility, crucial fixes for data averaging views. This
patch is a smallish refactor of the setup/dispatch/get call chain that
will permit the next patches.

The patch moves the reading of the GET variables into the conf data
structure at the setup() stage, sort of, assuming that all requests are
GET requests, which is currently true.