views: Allow request with comma-separated value lists in the query string
This fixes an issue with list values in hidden form elements.

Before, via Django, submitting list values in the query string was
limited to use of repeated keywords as in:

    /path/?foo=bar&foo=baz

now we also recognize the equivalent form:

    /path/?foo=bar,baz

Such GET request can be generated by GET forms with hidden elements that
code form list values.  The patch is limited in that any values must not
contain commas themselves as those will be parsed as list element
separators.