templates: fix base template for error pages
The patch populates the context via context processor for every request.
This is a bit sketchy of course but seems to be the simplest solution
while keeping the default error views. A proper solution would involve
writing our own error views using mibios.views.BaseMixin

Anyways with the patch the error pages look more like they belong to the
website.