models: refactor as_dataframe()
as_dataframe() now retrieves all field data in a single DB query.
Columns that need to display the natural value will incur an extra query
to retrtieve the id->natural mapping.

This fixes a crash in cases when the method is called to calculate the
sort-column statistics for some data with duplicate fields being
de-duplicated because ofthe implied distict() in the normal data
config's get_queryset().  That got the data column length out of wack
relative to the index and pandas raised (and luckily no bad results got
displayed).

It remains to see if this will perform good enough.