models: add progress metering for m2m relation setting
The implementation of the batching and progress metering bulk_create is
moved into a wrapper function.  The wrapper can be used when overriding
the manager's bulk_create or (as static method) can be wrapped around
some third manager's bulk_create as now done with the m2m relation add
during data import.