models: add loader managers for many umrad models
This commit re-organizes many models (except for UniRef100) loading code
into loader managers.  The LoadMixin remains for now to support
UniRef100.  The RefDBEntry model is now FuncRefDBEntry to better reflect
that it hosts references to functional databases.  It also has a loader
now that loads the data from the FunctionNames source table.