managers: Fix class inheritance issue with bulk_create() wrapping
Realized that bulk_create() comes from QuerySet and not BaseManager, so
we have to override it in a class derived from QuerySet.  And use the
from_queryset() mechanism to get the Loader class etc.

We want the wrapper to be usable outside QuerySet to wrap
other/third-party bulk_create methods, in managers for which we don't
control the class definitions.

Note that for umrad models the bulk_create() method of the regular
objects manager is also wrapped.