models: fix dataframe conversion for empty querysets
The values_list transposition code does not handle empty input lists
correctly, returning an empty list (since the second dimension is not
visible) when it should be a list of empty lists.

The bug became visible as crashed (table interface) web pages when
trying to sort on a column on an empty search result, or other empty
table.