views: add more features to table view plugins
While the plugins are not view objects, they can be used (a bit) like
them:

  * (new) setup() method, similar to View.setup()
  * (new) get_queryset(), takes the view's queryset as argument and is
    expected to return the final queryset.
  * (no changes) get_context_data(), provides template context as
    argument, updated context needs to be returned.

This patch improved how the plugins are procesed by TableView.  Future
Developments: if we want to extent this mechanism, it seems a better way
is to properly use model-specific TableView-derived classes (with the
model claqss attribute set appropriately), instead of this plugin
system.  The url conf then has to call a dispatcher to figure out which
view class to use.