glamr views: humanize key choices for Q filter editor
This adds a better version of the mibios.Model.get_related_accessors()
API.  The new get_related_accessors2() will not list id/natural fields.
It is based on the new get_related_fields() that lists the paths of fields
themselves.  The relation traversal algorithm is improved and avoids
going back one-to-one relations.