manager: in Loader.load() detect if create_from_m2m_input() can be called
create_from_m2m_input() only works for models that can be created from
an accession/identifier alone.  In other, more complex, cases we depend
on the referenced objects to already exist.  But that may not be the
case in certain circumstances, e.g. during testing if we only read in a
part of an input file.  If our m2m references refer to records of the
input file itself, then referenced records may not be listed in the
read-in portion of the file, and therefor missing at the point we want
to tor an m2m relation.  Hence, this commit will also include a check
that allows load() to continue, skipping saving those relation with the
target missing.