manager: Add infrastructure to catch bad input lines
This will in fact do very little.  Most bad data will only be caught at
the point of generating or submitting the SQL.  At that point we
usually, with the Loader.load() approach, have already read the whole
input file and lost any line number information.

This commit allows us to catch errors in conversion functions as those
run on each line individually.  The new InputFileError must be raised
inside the conversion function.  A message is printed and the line
skipped.  We permit a maximum of 10 errors (per file) to be treated like
this but for any further errors let the original exception propagate.

Some non-numerical content in the charge field in compound data file
will be caught this way.  This catches a bad line in the current
compound file.