managers: fix Loader setup
Using contribute_to_class() to set up the loader's
spec is not workable: the instance of Loader that will be used
eventually is different from the two or so (really!) available via
contribute_to_class.  The metaclass magic going on is quite involved and
the real loader instance not available until after the model is set up:
Model._meta.managers will at first access make a copy of the instance
added to Model._meta.local_managers during contribute_to_class and
<model>.<manager name> is a descriptor thingie.

So with this commit we simply do the setup at the beginning of
Loader.load().  Classes inheriting from Loader may want to be aware that
the spec won't be available before calling load())