utils: fix the atomic_dry wrapper to properly work with nested calls
If atomic_dry is called inside an already atomic_dry-wrapped function,
then the outer call should consume a dry_run=True parameter and do the
rollback.  The inner call should run with the default dry_run kwarg and
would normally not also do a rollback (unless dry_run is True by
default).  If dry_run as passed to the outer wrapper is not True, then
the parameter is passed on to the wrapped (outer) function and available
to be passed on to the inner wrapper and function.