utils: Redesign the timer facilities for the progress meter
It turns out there is a rather simple way to build the restarting
mechanics into the Timer class itself.  This nicely reduces the amount
and complexity of the code and (should) allow us to get rid of the extra
lock.

Adds some mechanism for the Loader to stop any active timer in case an
exception is now caught in _load_lines().