mibios_omics: drop support for split DBs
the split sqlite3 DB setup became to unwieldy, upcoming changes to the
search mechanism (using generic FKs) will require the contenttype table
to be in the same DB as the omics data.  Juggling this will be to
cumbersome and not worth any advantage of the split setup, if there ever
was one.