models: metagenomic pipeline integration work -- part II
This is the follow-up to commit 621b7367

With this commit series we stop using the results of the
AnnotateContigs.pl script (previously also known as Analyzecontigs.pl)
and directly consume the metagenomic analysis pipeline's output.  These
are the same input that Annotatecontig.pl consumes.  We try to replicate
a subset of it's processing steps.

Migration files are not part of this commit, there will be an aggregated
migration in the future.

With this update we load contig and genes their abundance and gene hits
to UniRef100.  LCAs can be calculated and an inititial calculation of
taxon abundance is implemented.

Not implemented yet are the derived compound and functional abundances.
The doc string at the top of the module has the as-of-now implemented
and working work-flow.