loader: Allow non-row / calculate value lines in InputFileSpec
Calculate a field value spec lines look like:

(InputFileSpec.CALC_VALUE, 'fieldname', spec_method)

These are processed after the row's regular data.  It allows the setting
of values of arbitrary fields.  Setting fields of related objects has
not been tested.  To allow more flexibility, with this patch, the (value
conversion) spec methods now pass along the object being processed in
addition to the row's data.  It should be noted that the object's fields
have not been cleaned yet, so the value of this is somewhat limited.