omics/models/managers: re-calculate rpkm/fpkm numbers
BBmap uses the count of reads that successfully map to normalize the
rpkm/fpkm numbers, that is the "m" in rpkm.  We find it more appropriate
to normalize by the sequencing effort.  With this commit we will use the
sample's number of post-QC read-pairs as a proxy for sequencing effort
and use it to re-calculate the rpkm/fpkm numbers.  The numbers
calculated by BBmap will be stored in the new rpkm_bbmap/fpkm_bbmap fields.

The migration that comes with the Contig/Gene model changes is not part
of this commit, it will be folded into a larger migration later.