umrad/loader: fix regression from introducing the CALC_VALUE mechanism
Adding the value-calculation-only file spec entries had some fallout,
some of our input file stopped loading.  In the end
InputfileSpec.setup()'s implementation was overhauled and the spec's
instance attributes re-designed.

As a result the row length is not checked anymore for each row.  Also,
for input files with headers, the order of items in the spec determines
the order of processing, while the input file may have any column
ordering.