glamr: sample meta data import: accept partial dates
Provide a solution for managing partial timestamps:
  * adds an extra field (collection_ts_partial) that marks which portion
    of the timestamp is actually provided
  * provide fake values for anythying missing and store as a regular datetime
  * assume the configured default timezone if TZ info is missing
  * so:
    - date only values get a fake time at midnight
    - day missing, set 1st of month @midnight
    - year only: set midnight, January 1st

This allows us to accept any MIMARKS/MIXS conform input and use any
built-in date and time querying and sorting functions.  For display, in
order to avoid showing the fake values, the collection_ts_partial fields
needs to be consulted to determine what to hide.  This last thing is not
part of this commit.