glamr: Fix for loading Dataset data in update mode
The accession field logic failed as the field order got messed up during
model inheritance, so the accession field is now declared explicitly.

Also, for the general Loader.load() in update mode, the retrieval of
objects to to be updated is made more robust by guarding against
duplicate record accessions.  This will make it easier to pinpoint
similar bugs in the future.