umrad/Loader: fix import with "computed values" fields
The spec's datastructures were not set up correctly in some cases.  And
it is now allowed not to have a conversion/processing method attached to
a CALC_VALUE spec item.  In these cases the value should be computed
during the processing of some other column.  The item must still be
present in the spec, at least for imports in update mode -- it is what
adds the field to the list of update fields, other fields won't get
saved to the DB.