models: Change many CharField into TextField
This was triggered by our move towards Postgresql which enforces
max_length on CharField.  Going forward, only certain string fields that
contain internally generated data, mostly choice fields, retain the
CharField type.  All others, whose values come from the outside are
changed to TextField.  Some retain the now optional max_length
parameter, set to what we think is an upper length bound for the
majority of values.  This can then be used as a hint for display
purposes.  Django will use it for it's text input widgets for forms,
something we don't plan to make use of currently.  A TextField without
max_length parameter set, shall indicate that we expect to see many very
long (meaning longer than a line of text on the screen) values.

This should have very little performance or storage space impact on
either Sqlite3 or Postgresql.

AccessionField and it's use remains unchanged, still evaluating this as
it's based on CharField and is populated with outside values of
ultimately unpredictable maximum length.

The corresponding migration will be committed later.